Annoyingly the WiFi portion of the ESP32 sometimes seems to drop out, probably because the radio is spending too much time on BT processing. This might need to be rewritten with ESP-IDF rather than the arduino wrappers.  Hmm.
When this happens loop() seems to abort quietly when trying to do an
MQTT reconnect.  A simple if{}else{} structure fails!

In the mean time I've put code before the reconnenct attempt; we just
try to connect 30 times (1 per second) and if it's still down then we
reboot.

If the MQTT server itself dies this does mean we reboot for no good
reason, but it should be harmless.